Create a FileSystemLayout class
